[Forward+] Reflection Probe Atlas Fallback
The Reflection Probe Atlas in Forward+ path allows access to "all" reflection probes in scene during path tracing, which improves visual quality of fallback areas.

The script attached to main camera is no longer required if URP Forward+ is the active rendering path.

Note: If there is no reflection probe in an area, the fallback data will be pure black in Forward+ path, which means the scene should be darker and less stable.